requirements.txt
discord
python-dotenv
giphy_client 